+++ {{cookiecutter.project_dirname}}/src/components/Navbar.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 
 import { ThemeSwitch } from '@/components/ThemeSwitch'
 import Image from 'next/image'
-import { useAppSelector } from '@/store'
+import { useAppSelector } from '@/src/store'
 
 const Navbar = () => {
   const theme = useAppSelector(state => state.utils.theme)
