+++ {{cookiecutter.project_dirname}}/src/components/ThemeSwitch.tsx
@@ -3,8 +3,8 @@ import styled from 'styled-components'
 import nookies from 'nookies'
 
 import { Theme } from '@/models/Utils'
-import { changeTheme } from '@/store/utilsSlice'
-import { useAppDispatch, useAppSelector } from '@/store'
+import { changeTheme } from '@/src/store/utilsSlice'
+import { useAppDispatch, useAppSelector } from '@/src/store'
 
 const ThemeSwitch = () => {
   const dispatch = useAppDispatch()
