+++ {{cookiecutter.project_dirname}}/src/pages/_app.tsx
@@ -2,7 +2,7 @@ import { Provider } from 'react-redux'
 import Head from 'next/head'
 import React from 'react'
 
-import store from '@/store'
+import store from '@/src/store'
 
 import Layout from '@/components/layout/Base'
