+++ {{cookiecutter.project_dirname}}/src/pages/api/health.ts
