+++ {{cookiecutter.project_dirname}}/src/store/utilsSlice.ts
