+++ {{cookiecutter.project_dirname}}/src/styles/themes/dark.ts
