+++ {{cookiecutter.project_dirname}}/src/styles/themes/index.ts
