+++ {{cookiecutter.project_dirname}}/src/styles/themes/light.ts
