+++ {{cookiecutter.project_dirname}}/src/utils/api/auth.ts
